Add owners functionality
Add a ownable interface to indicate that an entity can be owned by a
collection of owners.
Add a trait to easily implement the case of an entity with a single owner case.
Add a voter to check if a user has crud permissions based on ownership.
Add a subscriber that assigns ownership when adding/updating an entity
and the entity does not yet have an owner.