Add content negotation
Replace the controller hardcoded responses with the content negotion
bundle. This bundle will capture the generation of the response and
uses content negotation to determine which type of response is returned.
For instance, when the user agent accepts json, a json response with
content serialized with json is returned.

This enables us to reuse existing controller functions for an API.

Added some negotatiated content responses for existing calls and added
message descriptions for the api platform.

| Production Changes                                  | From | To     | Compare |
|-----------------------------------------------------|------|--------|---------|
| friendsofsymfony/rest-bundle                        | NEW  | 2.7.1  |         |
| gisostallenberg/response-content-negotiation-bundle | NEW  | 0.4.0  |         |
| symfony/contracts                                   | NEW  | v1.1.0 |         |
| willdurand/jsonp-callback-validator                 | NEW  | v1.1.0 |         |
| willdurand/negotiation                              | NEW  | v2.3.1 |         |